beats_f/requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Django==5.0.4
docutils==0.21.2
et-xmlfile==1.1.0
excel==1.0.1
fonttools==4.51.0
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-speech==2.26.0
google-generativeai==0.6.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
importlib_metadata==7.1.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jdcal==1.4.1
keyring==25.2.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplot==0.1.9
matplotlib==3.9.0
mdurl==0.1.2
more-itertools==10.2.0
mysqlclient==2.2.4
nh3==0.2.17
numpy==1.26.4
oauthlib==3.2.2
openpyxl==2.5.3
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pkginfo==1.10.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyloco==0.0.139
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32-ctypes==0.2.2
readme_renderer==43.0
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rsa==4.9
SimpleWebSocketServer==0.1.2
six==1.16.0
SpeechRecognition==3.10.4
sqlparse==0.5.0
tqdm==4.66.4
twine==5.1.0
typing==3.7.4.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
ushlex==0.99.1
websocket-client==1.8.0
whitenoise==6.6.0
XlsxWriter==3.2.0
xlwt==1.3.0
zipp==3.18.2
